server/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
